Clean up and move helper code to reusable spots
